[Chore/#7] move to request package
