+++ src/main/java/org/sopt/dto/Request/PostRequest.java
@@ -1,4 +1,4 @@
-package org.sopt.dto;
+package org.sopt.dto.Request;
 
 public record PostRequest(String title) {
 }
